+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 # [5.2.1]
 
 - rename cladding_offsets as bbox_offsets
+- copy_child_info propagates polarization and wavelength info
 
 ## [5.2.0]